Added: Limiter3

+++ Limiter/Source/PluginEditor.h
@@ -1,3 +1,20 @@
+/*
+ * Copyright (C) 2025 Filip Cenzak (filip.c@centrum.cz)
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program. If not, see <https://www.gnu.org/licenses/>.
+ */
+
 #pragma once
 
 #include <JuceHeader.h>
@@ -15,7 +32,7 @@ public:
 	// GUI setup
 	static const int N_SLIDERS = 5;
 	static const int SLIDERS[];
-	static const int COLUMN_OFFSET[];
+	static const float COLUMN_OFFSET[];
 	static const int N_ROWS = 1;
 	
 	//==============================================================================
